// src/models/apflora/Ekzaehleinheit.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { ApId } from './Ap';
import type { TpopkontrzaehlEinheitWerteId } from './TpopkontrzaehlEinheitWerte';

/** Identifier type for apflora.ekzaehleinheit */
export type EkzaehleinheitId = string & { __brand: 'apflora.ekzaehleinheit' };

/** Represents the table apflora.ekzaehleinheit */
export default interface Ekzaehleinheit {
  /** Primärschlüssel */
  id: EkzaehleinheitId;

  /** Zugehöriger Aktionsplan. Fremdschlüssel aus der Tabelle "ap" */
  ap_id: ApId;

  /** Zugehörige Zähleinheit. Fremdschlüssel aus der Tabelle "tpopkontrzaehl_einheit_werte" */
  zaehleinheit_id: TpopkontrzaehlEinheitWerteId | null;

  /** Ob die Zähleinheit zielrelevant ist */
  zielrelevant: boolean | null;

  /** Um die Zähleinheiten untereinander zu sortieren */
  sort: number | null;

  /** Bemerkungen zur EK-Zähleinheit */
  bemerkungen: string | null;

  /** Wer hat den Datensatz zuletzt geändert? */
  changed_by: string | null;

  /** Deklariert, dass bewusst keine der zwei Zähleinheiten von Massnahmen gewählt wurde. Ermöglicht, dass eine Qualitätskontrolle auflistet, wo unbewusst Zieleinheiten gewählt wurden, welche keiner der zwei Zähleinheiten von Massnahmen entsprechen */
  not_massn_count_unit: boolean | null;

  created_at: Date;

  updated_at: Date;
}

/** Represents the initializer for the table apflora.ekzaehleinheit */
export interface EkzaehleinheitInitializer {
  /**
   * Primärschlüssel
   * Default value: uuid_generate_v1mc()
   */
  id?: EkzaehleinheitId;

  /** Zugehöriger Aktionsplan. Fremdschlüssel aus der Tabelle "ap" */
  ap_id: ApId;

  /** Zugehörige Zähleinheit. Fremdschlüssel aus der Tabelle "tpopkontrzaehl_einheit_werte" */
  zaehleinheit_id?: TpopkontrzaehlEinheitWerteId | null;

  /**
   * Ob die Zähleinheit zielrelevant ist
   * Default value: false
   */
  zielrelevant?: boolean | null;

  /** Um die Zähleinheiten untereinander zu sortieren */
  sort?: number | null;

  /** Bemerkungen zur EK-Zähleinheit */
  bemerkungen?: string | null;

  /**
   * Wer hat den Datensatz zuletzt geändert?
   * Default value: NULL::character varying
   */
  changed_by?: string | null;

  /**
   * Deklariert, dass bewusst keine der zwei Zähleinheiten von Massnahmen gewählt wurde. Ermöglicht, dass eine Qualitätskontrolle auflistet, wo unbewusst Zieleinheiten gewählt wurden, welche keiner der zwei Zähleinheiten von Massnahmen entsprechen
   * Default value: false
   */
  not_massn_count_unit?: boolean | null;

  /** Default value: now() */
  created_at?: Date;

  /** Default value: now() */
  updated_at?: Date;
}

/** Represents the mutator for the table apflora.ekzaehleinheit */
export interface EkzaehleinheitMutator {
  /** Primärschlüssel */
  id?: EkzaehleinheitId;

  /** Zugehöriger Aktionsplan. Fremdschlüssel aus der Tabelle "ap" */
  ap_id?: ApId;

  /** Zugehörige Zähleinheit. Fremdschlüssel aus der Tabelle "tpopkontrzaehl_einheit_werte" */
  zaehleinheit_id?: TpopkontrzaehlEinheitWerteId | null;

  /** Ob die Zähleinheit zielrelevant ist */
  zielrelevant?: boolean | null;

  /** Um die Zähleinheiten untereinander zu sortieren */
  sort?: number | null;

  /** Bemerkungen zur EK-Zähleinheit */
  bemerkungen?: string | null;

  /** Wer hat den Datensatz zuletzt geändert? */
  changed_by?: string | null;

  /** Deklariert, dass bewusst keine der zwei Zähleinheiten von Massnahmen gewählt wurde. Ermöglicht, dass eine Qualitätskontrolle auflistet, wo unbewusst Zieleinheiten gewählt wurden, welche keiner der zwei Zähleinheiten von Massnahmen entsprechen */
  not_massn_count_unit?: boolean | null;

  created_at?: Date;

  updated_at?: Date;
}